tests/wpt/web-platform-tests/2dcontext/line-styles/2d.line.width.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.line.width.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.width.invalid</h1>
<p class="desc">Setting lineWidth to invalid values is ignored</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Setting lineWidth to invalid values is ignored");
_addTest(function(canvas, ctx) {

ctx.lineWidth = 1.5;
_assertSame(ctx.lineWidth, 1.5, "ctx.lineWidth", "1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = 0;
_assertSame(ctx.lineWidth, 1.5, "ctx.lineWidth", "1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = -1;
_assertSame(ctx.lineWidth, 1.5, "ctx.lineWidth", "1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = Infinity;
_assertSame(ctx.lineWidth, 1.5, "ctx.lineWidth", "1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = -Infinity;
_assertSame(ctx.lineWidth, 1.5, "ctx.lineWidth", "1.5");

ctx.lineWidth = 1.5;
ctx.lineWidth = NaN;
_assertSame(ctx.lineWidth, 1.5, "ctx.lineWidth", "1.5");


});
</script>
